backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
anyio==3.7.1
appdirs==1.4.4
argilla==1.19.0
async-timeout==4.0.3
asyncpg==0.27.0
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.11.0
boto3==1.28.46
botocore==1.31.46
cachetools==5.3.2
celery==5.2.7
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-log==0.4.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
cssselect==1.2.0
dataclasses-json==0.6.3
Deprecated==1.2.14
deprecation==2.1.0
docutils==0.20.1
docx2txt==0.8
dotty-dict==1.3.1
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.95.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
flake8==6.0.0
flake8-black==0.3.6
flower==2.0.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.36
google-ai-generativelanguage==0.4.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.0
googleapis-common-protos==1.61.0
gotrue==1.3.1
greenlet==3.0.1
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.19.4
humanize==4.9.0
idna==3.6
importlib-metadata==6.8.0
invoke==1.7.3
jaraco.classes==3.3.0
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.0
kombu==5.3.4
langchain==0.1.17
#langchain==0.0.304
#langchain_openai==0.0.6
langsmith==0.1.54
#langsmith==0.0.87
langgraph==0.0.45
litellm==0.1.816
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
newspaper3k==0.2.8
nh3==0.2.14
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openai==0.27.8
# openai==1.12.0
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
pathspec==0.11.2
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==4.0.0
postgrest==0.10.6
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.13
pyflakes==3.0.1
Pygments==2.17.2
PyMuPDF==1.22.3
pypandoc==1.11
pyparsing==3.1.1
pypdf==3.9.0
pyright==1.1.316
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-gitlab==3.15.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
python-semantic-release==7.33.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
readme-renderer==42.0
realtime==1.0.0
redis==4.5.4
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resend==0.5.1
rfc3986==1.5.0
rich==13.7.0
rsa==4.9
s3transfer==0.6.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.26.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.19
starlette==0.27.0
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformers==4.30.1
twine==3.8.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.6.7
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.12
webencodings==0.5.1
websockets==10.4
wrapt==1.14.1
xlrd==1.0.0
XlsxWriter==3.1.9
yarl==1.9.3
zipp==3.17.0
